Created PlacesFormPage
