+++ client/src/App.jsx
@@ -5,11 +5,14 @@ import LoginPage from './pages/LoginPage.jsx';
 import Layout from './Layout';
 import RegisterPage from './pages/RegisterPage';
 import axios from "axios";
-import AccountPage from './pages/AccountPage';
+import { UserContextProvider } from './UserContext';
+import ProfilePage from './pages/ProfilePage.jsx';
+import PlacesPage from './pages/PlacesPage';
+import PlacesFormPage from './pages/PlacesFormPage';
 
 axios.defaults.baseURL = 'http://127.0.0.1:4000';
 axios.defaults.withCredentials = true;
-import {UserContextProvider} from "./UserContext";
+
 
 function App() {
   return (
@@ -19,8 +22,9 @@ function App() {
         <Route index element={<IndexPage />} />
         <Route path="/login" element={<LoginPage />} />
         <Route path="/register" element={<RegisterPage />} />
-        <Route path="/account/:subPage?" element={<AccountPage />} />
-        <Route path="/account/:subPage/:action" element={<AccountPage />} />
+        <Route path="/account" element={<ProfilePage />} />
+        <Route path="/account/places" element={<PlacesPage />} />
+        <Route path="/account/places/new" element={<PlacesFormPage />} />
         </Route>
       </Routes>
     </UserContextProvider>
@@ -29,4 +33,4 @@ function App() {
 
 export default App
 
-//Dan - left off at 1:48:00
+
