+++ client/src/pages/PlacesPage.jsx
@@ -1,54 +1,15 @@
-import Perks from "../perks.jsx";
-import { Link, Navigate } from "react-router-dom";
-import axios from "axios";
-import PhotosUploader from "../PhotosUploader.jsx";
-
+import { Link, useParams } from "react-router-dom";
+import AccountNav from "../AccountNav";
 export default function PlacesPage() {
-  const {action} = useParams();
-  const [title,setTitle] = useState('');
-  const [address,setAddress] = useState('');
-  const [addedPhotos,setAddedPhotos] = useState([]);
-  const [description,setDescription] = useState('');
-  const [perks,setPerks] = useState('');
-  const [extraInfo,setExtraInfo] = useState('');
-  const [checkIn,setCheckIn] = useState('');
-  const [checkOut,setCheckOut] = useState('');
-  const [maxGuests,setMaxGuests] = useState('');
-  const [redirectToPlacesList, setRedirectToPlacesList] = useState(false);
-  function inputHeader(text) {
-   return (
-    <h2 className="text-2xl mt-4">{text}</h2>
-   );
-  }
-  function inputDescription(text) {
-    return(
-      <p className="test-gray-500 text-sm">{text}</p>
-    )
-  };
-  function preInput(header,description) {
-    return(
-      <>
-      {inputHeader(header)}
-       {inputDescription(description)}
-      </>
-    );
-  }
-
-  async function addNewPlace(ev){
-    ev.preventDefault()
-    await axios.post('/places', {
-      title, address, addedPhotos, 
-      description, perks, extraInfo, 
-      checkIn, checkOut, maxGuests
-    })
-    setRedirectToPlacesList(true)
-  }
-  if(redirectToPlacesList && !action=='new') {
-    return <Navigate to={'/account/places'} />
-  }
+  const [places, setPlaces] = useSate([])
+ useEffect(() => {
+  axios.get('/places').then(({data}) => {
+    setPlaces(data)
+  })
+ }, []);
 return (
     <div>
-      {action !== 'new' && (
+      <AccountNav />
         <div className="text-center">
           <Link className="inline-flex gap-1bg-primary text-white py-2 px-6 rounded-full" to ={'/account/places/new'}>
             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
@@ -56,52 +17,14 @@ return (
             </svg>
           Add new place
           </Link>
-        </div>
-)}
-    {action === 'new' && (
-      <div>
-        <form onSubmit={addNewPlace}>
-          {preInput('Title','Title for your location, should be unique.')}
-          <input type="text" value={title} onChange={ev => setTitle(ev.target.value)} placeholder="title, example: paris paradise!" />
-          {preInput('Address','Address to Location')}
-          <input type="text" value={address} onChange={ev => setAddress(ev.target.value)} placeholder="address" />
-          {preInput('Photos','more = better')}
-          <PhotosUploader />
-          {preInput('Description','description of location')}
-          <textarea value={description} onChange={ev => setDescription(ev.target.value)}/>
-              {preInput('Perks','Select any perks that make your spot a breeze')}
-          <div className="grid mt-2 gap-2 grid-cols-2 md:grid-cols-4 lg:grid-cols-6">
-            <Perks selected={perks} onChange={setPerks} />
-          </div>
-          {preInput('extra information','House rules and guidelines')}
-          <textarea value={extraInfo} onChange={ev => setExtraInfo(ev.target.value)} />
-          {preInput('check-in & checkout times, max guests','check in&out times should include time for cleaning between breeze guests.')}
-          <div className="grid gap-2 sm:grid-cols-3 ">
-            <div>
-              <h3 className="mt-2 -mb-1">Check-in time:</h3>
-              <input type="text" value={checkIn} onChange={ev => setCheckIn(ev.target.value)} placeholder="8:00am"/>
-              </div>
-              <div>
-              <h3 className="mt-2 -mb-1">Checkout time:</h3>
-              <input type="text" 
-              value ={checkOut} 
-              onChange={ev => setCheckOut(ev.target.value)} 
-              placeholder="3:00pm"/>
-              </div>
-              <div>
-              <h3 className="mt-2 -mb-1">Max guests:</h3>
-              <input type="number" 
-              value={maxGuests} 
-              onChange={ev => setMaxGuests(ev.target.value)}/> 
-            </div>
-          </div>
           <div>
-            <button className="bg-primary my-4 text-white py-2 px-6 rounded-full">Save</button>
+            {places.length > 0 && places.map(place => (
+              <div>
+                {place.title}
+              </div>
+            ))}
           </div>
-        </form>
-      </div>
-    )}
-    my places
+        </div>
   </div>
 )
 }
